+++ examples/framesjs-starter/app/debug/components/create-signer.tsx
@@ -16,7 +16,7 @@ export const LoginWindow = ({
 }) => {
   return (
     <div>
-      <div style={{ minWidth: "150px" }}>
+      <div style={{ minWidth: "150px" }} className="mt-4">
         <div>
           {farcasterUser?.status === "approved" ? (
             farcasterUser.fid ? (
@@ -29,10 +29,16 @@ export const LoginWindow = ({
             )
           ) : farcasterUser?.status === "impersonating" ? (
             <div>
-              Impersonating fid (for testing, only works for local frames using
-              frames.js to validate messages, as they&apos;re mocked):
-              <b>{farcasterUser?.fid}</b>{" "}
-              <button onClick={logout}>Logout</button>
+              Impersonating fid: <b>{farcasterUser?.fid}</b>,{" "}
+              <button className="underline" onClick={logout}>
+                Logout
+              </button>
+              <p>
+                <span className=" text-slate-400">
+                  *Impersonation only works for testing local frames using
+                  frames.js to validate messages, as they&apos;re mocked
+                </span>{" "}
+              </p>
             </div>
           ) : farcasterUser?.status === "pending_approval" ? (
             "Approve in Warpcast"