+++ packages/frames.js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # frames.js
 
+## 0.3.0
+
+### Minor Changes
+
+- 03d041e: feat: mint action
+
+## 0.2.1
+
+### Patch Changes
+
+- 0d9cfaf: Fix validation for order
+
 ## 0.2.0
 
 ### Minor Changes